+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+# Version 1.6
+**Added**
+ - A Game Rule config to disallow people who join your lobby from using the hotbar keybinds
+ - A Game Rule config to disallow people from accessing the terminal
+ - A Game Rule config to use vanilla defaults for all Sprint related options without having to manually change each one to the vanilla default (does not overwrite your stored values, only in-game values use the vanilla ones instead)
+ - A Game Rule config to use vanilla defaults for all Stamina related options without having to manually change each one to the vanilla default (does not overwrite your stored values, only in-game values use the vanilla ones instead)
+ - A Game Rule config to use vanilla defaults for all Terminal Item Weight options without having to manually change each one to the vanilla default (does not overwrite your stored values, only in-game values use the vanilla ones instead)
+ - If you have any of the vanilla override game rules set to true, when you join someone's lobby you will use the vanilla default settings for that category (stamina/sprint speed/item weights), instead of use the synced ones from the host.
+
 # Version 1.5
 ## IMPORTANT!
  - I've renamed a few configs so it's best to delete the config file you have and let it regenerate by launching the game and change the configs back to your liking.
 
 **Added**
+ - BepinEx as dependency to be installed.
  - A config that allows you to bind the first available flashlight in your inventory to a key
- - A config that allows you to dissallow people from using the keybind if they join your lobby
+ - A config that allows you to disallow people from using the flashlight keybind if they join your lobby
  - Configs that allow you to modify Stamina Recharging
 
 # Version 1.4.3